fix: make the get-object return a commit if looking for a commit
This makes it act like an identity function, which can be helpful
rather than forcing users to convert a commit to an oid to have it
looked up as a commit again later on.

e.g.

  (git:make-commit (index-to-tree idx)
                   "message"
                   :author author
                   :repository repo
                   :parents (git:resolve (git:repository-head repo)))

In this example the parent would be a commit object, but the function
expects an oid, which it then looks up as a commit.